增加流量统计接口，后续还需要启用定时器，每小时更新日、周、月、季度、年流量

+++ ssserver/src/main/java/com/hellozjf/shadowsocks/ssserver/dataobject/FlowStatisticsDetail.java
@@ -22,7 +22,7 @@ public class FlowStatisticsDetail extends BaseEntity {
     /**
      * 入/出站的流量
      */
-    private Integer flowSize;
+    private Long flowSize;
 
     /**
      * Client所访问的SS服务端地址
